fix: correct flow writer cancel method to use cancel instead of commit
There seems to be a logical error in mxlFlowWriterCancelSamples(),
it invokes commit() on the writer, which is committing the samples
rather than discarding them. This means that calls to cancel a write
actually ended up committing the data.

This patch changes the implementation to call cancel() instead of
commit(), restoring the intended behaviour for cancelled samples.

Signed-off-by: moleksy <maciej.oleksy@intel.com>