Change how partial grains work. Check shm struct versions
Change how partial grain support works in mxl. Now, instead of allowing
producers to choose freely how many bytes they want to commit to a grain, we
dictate a slice length. The progress of writing to a grain is signalled
through a number of these slices that have been written to the grain. A
slice of v210 video is equivalent to a picture line. A slice of "data"
is a single byte.

Also adds a check for the correct struct version number whenever a
shared memory segment is opened and interpreted as a FlowInfo or
GrainInfo struct.

Co-authored-by: Jonas Ohland <jonas.ohland@gmail.com>
Signed-off-by: Jonas Ohland (Riedel) <jonas.ohland@riedel.net>